Update llvm-9999.ebuild
	This is a personal set of fixes pulled from changes made in
/etc/portage/env/sys-devel/llvm-9999 to allow llvm-9999 to compile on my
system.  I don't believe there is anything tremendously unusual about my
build environment, and most of the issues seemed to be primarily caused
by changes to the build system and source/binary/resource locations in
the LLVM and Clang/Compiler-RT Git master branch.

	If you are a Gentoo developer thinking about merging this into
repo/gentoo.git, *please* check it over and test it on your system.
This commit/patch may find its way into a bug, but I very likely will
not be submitting a pull request or mailbox patchset to get this
committed, because it's only been tested on one system, and while credit
is always nice, I do *not* want my name attached to a commit that sets
users' computers on fire in an append-only SCM log.